Site ajustado agora aparecendo cardapio

+++ meu_app/models.py
@@ -31,15 +31,15 @@ class Questionario(models.Model):
     usuario = models.ForeignKey(User, on_delete=models.CASCADE, related_name="questionarios", verbose_name="Usuário")
     nome = models.CharField(max_length=100, blank=True, null=True)
     idade = models.IntegerField(null=True, blank=True)
-    peso = models.FloatField(null=True, blank=True)  # em kg
-    altura = models.FloatField(null=True, blank=True)  # em metros
+    peso = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)  # em kg
+    altura = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)  # em metros
     objetivo = models.CharField(max_length=100)
     restricoes = models.TextField(blank=True)
     preferencia = models.TextField(blank=True)
     refeicoes_por_dia = models.IntegerField(null=True, blank=True)
     come_carne = models.BooleanField(default=True)
     gosta_de_legumes = models.BooleanField(default=True)
-    agua_bebida = models.FloatField(default=0, null=True, blank=True)  # ml de água por dia
+    agua_bebida = models.DecimalField(max_digits=5, decimal_places=2, default=0, null=True, blank=True)  # ml de água por dia
     usa_suplementos = models.BooleanField(default=False)
     estresse = models.CharField(max_length=100, blank=True)
     atividade_fisica = models.CharField(max_length=100, choices=ATIVIDADE_FISICA_CHOICES, blank=True)
@@ -105,8 +105,8 @@ class Perfil(models.Model):
     data_nascimento = models.DateField(null=True, blank=True)
     genero = models.CharField(max_length=20, blank=True)
     endereco = models.CharField(max_length=255, blank=True)
-    peso = models.FloatField(null=True, blank=True)
-    altura = models.FloatField(null=True, blank=True)
+    peso = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
+    altura = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
     objetivo = models.CharField(max_length=100, blank=True)
     restricoes = models.TextField(blank=True)
     preferencia = models.TextField(blank=True)
